skonczony bmi

+++ src/Zadanie.java
@@ -4,7 +4,7 @@ public class Zadanie {
 	public static void main(String[] args) {
 		Scanner in = new Scanner(System.in);
 		int weight = in.nextInt();
-		int height = in.nextInt();
+		double height = in.nextDouble();
 		
 		
 		in.close();
